ok long went away in python3

+++ python/pyspark/ml/param/_shared_params_code_gen.py
@@ -124,7 +124,7 @@ if __name__ == "__main__":
         ("outputCol", "output column name.", "self.uid + '__output'", None),
         ("numFeatures", "number of features.", None, int),
         ("checkpointInterval", "checkpoint interval (>= 1).", None, int),
-        ("seed", "random seed.", "hash(type(self).__name__)", long),
+        ("seed", "random seed.", "hash(type(self).__name__)", int),
         ("tol", "the convergence tolerance for iterative algorithms.", None, float),
         ("stepSize", "Step size to be used for each iteration of optimization.", None, float),
         ("handleInvalid", "how to handle invalid entries. Options are skip (which will filter " +
